[INTERNAL_BRANCH=salazarm/catalog-update] feedback
